changed algorythm

+++ docker/gp_tests/scripts/tests/ao_storage_test.sh
@@ -17,10 +17,18 @@ enable_pitr_extension
 wal-g --config=${TMP_CONFIG} delete everything FORCE --confirm
 
 # 1st full backup (init tables heap, ao, co)
-insert_data
+psql -p 6000 -c "DROP DATABASE IF EXISTS test"
+psql -p 6000 -c "CREATE DATABASE test"
+psql -p 6000 -d test -c "CREATE TABLE ao(a int, b int) WITH (appendoptimized = true) DISTRIBUTED BY (a);"
+psql -p 6000 -d test -c "CREATE TABLE co(a int, b int) WITH (appendoptimized = true, orientation = column) DISTRIBUTED BY (a);"
 psql -p 6000 -d test -c "INSERT INTO co select i, i FROM generate_series(1,1000000)i;"
 psql -p 6000 -d test -c "INSERT INTO ao select i, i FROM generate_series(1,1000000)i;"
 run_backup_logged ${TMP_CONFIG} ${PGDATA} "--full"
+var=$(wal-g --config=${TMP_CONFIG} st ls segments_005/seg0/basebackups_005/aosegments/ | grep aoseg | wc -l)
+if [ "$var" -ne 3 ] ; then
+    echo "Error: expected 3 aoseg files but found $var"
+    exit 1
+fi
 
 # 2nd backup (populate the co table)
 psql -p 6000 -d test -c "DELETE FROM co WHERE a>0;"
@@ -29,7 +37,11 @@ psql -p 6000 -d test -c "VACUUM FULL;"
 psql -p 6000 -d test -c "INSERT INTO ao select i, i FROM generate_series(10000000,20000000)i;"
 psql -p 6000 -d test -c "INSERT INTO co select i, i FROM generate_series(10000000,20000000)i;"
 run_backup_logged ${TMP_CONFIG} ${PGDATA}
-
+var=$(wal-g --config=${TMP_CONFIG} st ls segments_005/seg0/basebackups_005/aosegments/ | grep aoseg | wc -l)
+if [ "$var" -ne 9 ] ; then
+    echo "Error: expected 9 aoseg files but found $var"
+    exit 1
+fi
 
 stop_and_delete_cluster_dir
